Add files via upload
The data mining algorithm applied in this submission is a multinomial Bayesian classifier. This algorithm makes an assumption that all the predictor variables in the data set are mutually independent (Chauhan, 2018). For the given data set, the predictor variables are words in Reddit submissions’ titles. As the article explains, the algorithm is based on the relationship between classes (Reddit subreddit names, in this assignment) and features (tfidf of words in the submission, in this assignment) defined by the Bayes theorem.  

Probability of a text with features x1...xn belonging to a class C is the probability of having these features given class C multiplied by the probability of belonging to class C divided by the probability of having the given features, as summarized in the article. 
 In the attached Python script, the submissions for two different subreddits were pre-processed for model fitting according to a tutorial in Smetanin (2018). The multinomial Bayesian classifier used in the tutorial estimates the probability of a feature i appearing in a sample that belongs to the class C as a smoothed version of relative frequency counting: the number of times feature i appears in a sample of class k in the training set divided by the total count of all features for class C. 

Chauhan, G. (2018) All about Naive Bayes, towards data science, Oct 8, 2018, available: https://towardsdatascience.com/all-about-naive-bayes-8e13cef044cf [accessed 6 April 2021]

Smetanin, S (2018) Sentiment Analysis of Tweets using Multinomial Naive Bayes, towards data science, Sep 1, available: https://towardsdatascience.com/sentiment-analysis-of-tweets-using-multinomial-naive-bayes-1009ed24276b  [accessed 6 April 2021]